+++ tests/builder/app_from_crate.rs
