+++ tests/builder/borrowed.rs
