+++ tests/builder/default_vals.rs
@@ -1,4 +1,4 @@
-mod utils;
+use crate::utils;
 use clap::{arg, App, Arg, ErrorKind};
 
 #[test]
