+++ tests/builder/double_require.rs
