+++ tests/builder/global_args.rs
@@ -1,5 +1,3 @@
-mod utils;
-
 use clap::{arg, App, Arg};
 
 #[test]
