+++ tests/builder/grouped_values.rs
@@ -1,7 +1,5 @@
 #![cfg(feature = "unstable-grouped")]
 
-mod utils;
-
 use clap::{App, AppSettings, Arg};
 
 #[test]
