+++ tests/builder/ignore_errors.rs
