+++ tests/builder/main.rs
@@ -1,0 +1,47 @@
+mod app_from_crate;
+mod app_settings;
+mod arg_aliases;
+mod arg_aliases_short;
+mod arg_matcher_assertions;
+mod arg_settings;
+mod borrowed;
+mod cargo;
+mod conflicts;
+mod default_missing_vals;
+mod default_vals;
+mod delimiters;
+mod derive_order;
+mod display_order;
+mod double_require;
+mod empty_values;
+mod env;
+mod error;
+mod flag_subcommands;
+mod flags;
+mod global_args;
+mod grouped_values;
+mod groups;
+mod help;
+mod help_env;
+mod hidden_args;
+mod ignore_errors;
+mod indices;
+mod multiple_occurrences;
+mod multiple_values;
+mod opts;
+mod positionals;
+mod posix_compatible;
+mod possible_values;
+mod propagate_globals;
+mod regex;
+mod require;
+mod subcommands;
+mod template_help;
+mod tests;
+mod unicode;
+mod unique_args;
+mod utf16;
+mod utf8;
+mod utils;
+mod validators;
+mod version;
