+++ tests/builder/multiple_occurrences.rs
